requirements.txt
flask
discord.py
requests
pyyaml
websockets
